practice_file/0206/0206_team2/Producer/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.2",
    "kafka-python>=2.3.0",
    "pydantic-settings>=2.12.0",
]
